+++ model/scratch/dffml_model_scratch/slr.py
@@ -15,7 +15,7 @@ from dffml.repo import Repo
 from dffml.base import config, field
 from dffml.source.source import Sources
 from dffml.feature import Features
-from dffml.accuracy import Accuracy
+from dffml.model.accuracy import Accuracy
 from dffml.model.model import ModelConfig, ModelContext, Model, ModelNotTrained
 from dffml.util.entrypoint import entrypoint
 from dffml.util.cli.arg import Arg
@@ -129,9 +129,11 @@ class SLRContext(ModelContext):
     ) -> AsyncIterator[Tuple[Repo, Any, float]]:
         if self.regression_line is None:
             raise ModelNotTrained("Train model before prediction.")
+        target = self.parent.config.predict.NAME
         async for repo in repos:
             feature_data = repo.features(self.features)
             repo.predicted(
+                target,
                 await self.predict_input(feature_data[self.features[0]]),
                 self.regression_line[2],
             )
@@ -187,8 +189,10 @@ class SLR(Model):
                 },
                 "last_updated": "2019-07-19T09:46:45Z",
                 "prediction": {
-                    "confidence": 1.0,
-                    "value": 90.0
+                    "Salary": {
+                        "confidence": 1.0,
+                        "value": 90.0
+                    }
                 },
                 "key": "0"
             }