+++ source/mysql/dffml_source_mysql/source.py
@@ -36,11 +36,22 @@ class MySQLSourceContext(BaseSourceContext):
                 key_value_pairs[modified_key] = repo.data.features[
                     modified_key
                 ]
-            elif key.startswith("prediction_"):
-                modified_key = key.replace("prediction_", "")
-                key_value_pairs[modified_key] = repo.data.prediction[
-                    modified_key
-                ]
+            elif "_value" in key:
+                target = key.replace("_value", "")
+                if repo.data.prediction:
+                    key_value_pairs[key] = repo.data.prediction[target][
+                        "value"
+                    ]
+                else:
+                    key_value_pairs[key] = "undetermined"
+            elif "_confidence" in key:
+                target = key.replace("_confidence", "")
+                if repo.data.prediction:
+                    key_value_pairs[key] = repo.data.prediction[target][
+                        "confidence"
+                    ]
+                else:
+                    key_value_pairs[key] = 1
             else:
                 key_value_pairs[key] = repo.data.__dict__[key]
         db = self.conn
@@ -51,19 +62,18 @@ class MySQLSourceContext(BaseSourceContext):
         self.logger.debug("update: %s", await self.repo(repo.key))
 
     def convert_to_repo(self, result):
-        modified_repo = {
-            "key": "",
-            "data": {"features": {}, "prediction": {}},
-        }
+        modified_repo = {"key": "", "data": {"features": {}, "prediction": {}}}
         for key, value in result.items():
             if key.startswith("feature_"):
                 modified_repo["data"]["features"][
                     key.replace("feature_", "")
                 ] = value
-            elif key.startswith("prediction_"):
-                modified_repo["data"]["prediction"][
-                    key.replace("prediction_", "")
-                ] = value
+            elif ("_value" in key) or ("_confidence" in key):
+                target = key.replace("_value", "").replace("_confidence", "")
+                modified_repo["data"]["prediction"][target] = {
+                    "value": result[target + "_value"],
+                    "confidence": result[target + "_confidence"],
+                }
             else:
                 modified_repo[key] = value
         return Repo(modified_repo["key"], data=modified_repo["data"])
@@ -81,22 +91,23 @@ class MySQLSourceContext(BaseSourceContext):
         db = self.conn
         await db.execute(query, (key,))
         row = await db.fetchone()
+
         if row is not None:
+            features = {}
+            predictions = {}
+            for key, value in row.items():
+                if key.startswith("feature_"):
+                    features[key.replace("feature_", "")] = value
+                elif "_value" in key:
+                    target = key.replace("_value", "")
+                    predictions[target] = {
+                        "value": row[target + "_value"],
+                        "confidence": row[target + "_confidence"],
+                    }
             repo.merge(
                 Repo(
                     row["key"],
-                    data={
-                        "features": {
-                            key.replace("feature_", ""): value
-                            for key, value in row.items()
-                            if key.startswith("feature_")
-                        },
-                        "prediction": {
-                            key.replace("prediction_", ""): value
-                            for key, value in row.items()
-                            if key.startswith("prediction_")
-                        },
-                    },
+                    data={"features": features, "prediction": predictions},
                 )
             )
         return repo