+++ tests/integration/test_service_dev.py
@@ -7,7 +7,6 @@ import json
 import inspect
 import pathlib
 import unittest.mock
-
 from dffml.df.types import DataFlow
 from dffml.cli.cli import CLI
 from dffml.service.dev import Develop